Polish Settings: one switch style and clearer sections.
Share SettingsSwitch/ToggleRow chrome, fold preferences into a Display stack with dividers, and separate Addresses / Wallet / danger zone.
